% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISEP_cv.R
\name{BISEP_cv}
\alias{BISEP_cv}
\title{BISEP_cv}
\usage{
BISEP_cv(data = data, seed_val = NULL)
}
\arguments{
\item{data}{This should be a log2 gene expression matrix with genes as rownames and samples as
column names. Suitable for gene expression data from any platform - NGS datasets should be
RPKM or RSEM values.}

\item{seed_val}{A numeric.  Default is NULL - seed not set.}
}
\value{
A list containing three matrices. Matrix 1 contains the output of the BISEP algorithm -
  including the midpoint of the bimodal distribution and the associated p value. Matrix 2
  contains the output from the BI algorithm - including the delta, pi and BI values.
  Matrix 3 contains the input matrix.
}
\description{
Input a matrix of continuous data with genes as rows and samples as columns.
Algorithm may take some considerable time to run (1 hour for 20,500 genes across 800 samples)
}
\details{
Note: this is a slightly modified version of the BISEP algorithm from the
 \code{\link[BiSEp]{BiSEp}} package: see the original \code{\link[BiSEp]{BISEP}} function documentation.
 The modification allows the random seed to be set so that p-values generated by sampling stay the same
 from run to run.
}
\examples{
library(BiSEp)
data("INPUT_data")
bisep_out <- BISEP_cv(INPUT_data, seed_val = 12345)
bisep_out$BISEP
bisep_out$BI
}
\author{
Mark Wappett, \email{m.a.wappett at googlemail.com}
}
\references{
\href{https://bmcgenomics.biomedcentral.com/articles/10.1186/s12864-016-2375-1}{Wappett et al BMC Genomics 2016}
}
\seealso{
\code{\link[BiSEp]{BiSEp}}
}

