% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineResults.R
\name{combineResults}
\alias{combineResults}
\title{combineResults}
\usage{
combineResults(con, genes)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database containing the data}

\item{genes}{A vector of ensembl gene ids}
}
\value{
a data frame
}
\description{
Extract and combine results from all analyses for all genes
}

