% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countHumanParalogs.R
\name{countHumanParalogs}
\alias{countHumanParalogs}
\title{countHumanParalogs}
\usage{
countHumanParalogs(con, genes)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}
}
\value{
a data frame
}
\description{
This function uses the \pkg{biomaRt} package to extract paralog information for supplied human genes
}

