% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBISEPAnalysis.R
\name{doBISEPAnalysis}
\alias{doBISEPAnalysis}
\title{BISEP Analysis}
\usage{
doBISEPAnalysis(con, genes_n = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{genes_n}{Permits a random subsample of genes to be analysed rather than the full set}
}
\value{
A data frame
}
\description{
The BISEP analysis function identifies genes with bimodel expression profiles.  See \code{\link[BiSEp]{BiSEp}} documentation
}

