% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMutationAnalysis.R
\name{doMutationAnalysis}
\alias{doMutationAnalysis}
\title{doMutationAnalysis}
\usage{
doMutationAnalysis(con, genes)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{genes}{A vector of human ENSEMBL gene ids}
}
\value{
a data frame with the results of the analysis
}
\description{
This function calculates the proportion of samples with either a loss of function mutation (truncation etc)
or just a general protein coding change mutation in the cohort
}

