% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRNAseqPlot.R
\name{doRNAseqPlot}
\alias{doRNAseqPlot}
\title{RNAseq plot}
\usage{
doRNAseqPlot(con, input_gene, table_name = "tcga_rnaseq_data")
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{input_gene}{A vector of one or more ENSEMBL gene ids}

\item{table_name}{The table in the SQLite database containing the data.  The table should have fields named
patient_id, gene_id and value where gene_id is an ensembl id.  Default tcga_rnaseq_data}
}
\value{
A ggplot2 object
}
\description{
View RNAseq data for a given gene
}

