% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doWormMineAnalysis.R
\name{doWormMineAnalysis}
\alias{doWormMineAnalysis}
\title{doWormMineAnalysis}
\usage{
doWormMineAnalysis(con, genes, alt_mart = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{genes}{A vector of ENSEMBL gene ids}

\item{alt_mart}{A Mart object created using biomaRt::useMart.  This should specify \code{hsapiens_gene_ensembl}
as the dataset.  See \code{\link{getOrthologs}} for more details.}
}
\value{
a data frame
}
\description{
This function determines the lethality status in C elegans of a given human gene.
}
\details{
Note: queries the WormMine API using Ruby see http://im-dev.wormbase.org/tools/wormmine/api.do?subtab=ruby
}

