% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrthologs.R
\name{getOrthologs}
\alias{getOrthologs}
\title{getOrthologs}
\usage{
getOrthologs(con, genes, species = "mmusculus", alt_mart = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{genes}{A vector of ENSEMBL gene ids}

\item{species}{The ENSEMBL species id eg mmusculus for mouse}

\item{alt_mart}{A Mart object created using biomaRt::useMart.  This should specify \code{hsapiens_gene_ensembl}
as the dataset.}
}
\value{
a data frame
}
\description{
This function uses the \pkg{biomaRt} package to get ortholog information for supplied human genes
}
\examples{
\dontrun{
ex_con <- setupSQLite()
ex_genes <- c('ENSG00000074800', 'ENSG00000127616')

#get mouse orthologs
getOrthologs(ex_con, ex_genes, species='mmusculus')

#get fly orthologs
getOrthologs(ex_con, ex_genes, species='dmelanogaster')

#view the tables in the database
DBI::dbListTables(ex_con)

#larger set of genes
all_genes <- getAllHumanGenes(ex_con)
set.seed(20120806)
ex_genes2 <- sample(all_genes$gene_id, 2000)
ortho_df <- getOrthologs(ex_con, ex_genes2, species='mmusculus')

#Specify which Mart to use
my_mart <- biomaRt::useMart("ENSEMBL_MART_ENSEMBL", dataset="hsapiens_gene_ensembl", host="sep2015.archive.ensembl.org")
getOrthologs(ex_con, ex_genes, species='mmusculus', alt_mart=my_mart)
getOrthologs(ex_con, ex_genes, species='dmelanogaster', alt_mart=my_mart)
}
}

