% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTCGAMutationData.R
\name{getTCGAMutationData}
\alias{getTCGAMutationData}
\title{Get TCGA mutation data}
\usage{
getTCGAMutationData(con, cancerTypes = "LUAD", releaseDate = "2015-11-01",
  sampletag = "01A", sampleids = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{releaseDate}{GDAC release date.  See \code{\link[RTCGA]{downloadTCGA}} documentation}

\item{sampletag}{Refers to the sample part of the barcode.  Default is 01A.  See \link[=https://wiki.nci.nih.gov/display/TCGA/TCGA+barcode]{TCGA wiki}}

\item{sampleids}{A vector of sample ids to include.  Default is NULL - ie no filtering.  See \link[=https://wiki.nci.nih.gov/display/TCGA/TCGA+barcode]{TCGA wiki}}
}
\value{
a data frame
}
\description{
Downloads, filters and populates the SQLite database with TCGA mutation data using functionality from the \code{\link[RTCGA]{RTCGA}} package
}

