% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTCGARNAseqData.R
\name{getTCGARNAseqData}
\alias{getTCGARNAseqData}
\title{Get TCGA RNAseq data}
\usage{
getTCGARNAseqData(con, releaseDate, sampletag = "01A", sampleids = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{sampleids}{}
}
\value{
a data frame
}
\description{
Downloads, filters and populates the SQLite database with TCGA RNAseq data using functionality from the \pkg{RTCGA}
}

