% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBISEPOutput.R
\name{plotBISEPOutput}
\alias{plotBISEPOutput}
\title{BISEP Plot}
\usage{
plotBISEPOutput(con, highlighted_gene = NULL, pi_value_th = 0.2,
  bisep_pval_th = 0.1, bi_value_th = 0, chromosome_sel = 0,
  xval = "pi_value", yval = "bisep_pval", colval = "selected")
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{highlighted_gene}{Highlight the provided gene on the plot.  Default NULL - ie don't highlight anything}

\item{pi_value_th}{Set a PI value threshold.  Default 0.2}

\item{bisep_pval_th}{Set a BISEP p-value threshold.  Default 0.1}

\item{bi_value_th}{Set a Bimodal Index threshold.  Default 0}

\item{chromosome_sel}{A vector of chromosome id's.  Default is 0 which means all chromosomes.}

\item{xval}{The variable to plot on the x-axis.}

\item{yval}{The variable to plot on the y-axis}
}
\value{
A ggplot2 object
}
\description{
Plot to visualise BISEP output
}

