% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCombinedResults.R
\name{plotCombinedResults}
\alias{plotCombinedResults}
\title{Plot of combined results}
\usage{
plotCombinedResults(con, highlighted_gene = NULL, pi_value_th = 0.2,
  bisep_pval_th = 0.1)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{highlighted_gene}{Highlight the provided gene on the plot.  Default NULL - ie don't highlight anything}

\item{pi_value_th}{Set a PI value threshold.  Default 0.2}

\item{bisep_pval_th}{Set a BISEP p-value threshold.  Default 0.1}
}
\value{
A ggplot2 object
}
\description{
Uses scatter3D package to generate an interactive scatter plot of the combined results
}

