% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParalogRNAseq.R
\name{plotParalogRNAseq}
\alias{plotParalogRNAseq}
\title{Paralog expression}
\usage{
plotParalogRNAseq(con, input_gene)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{gene_id}{An ensembl gene ID}
}
\value{
A ggplot2 object
}
\description{
Compare RNAseq data for a given gene vs its paralogs
}

