% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisApp.R
\name{shinyVisApp}
\alias{shinyVisApp}
\title{Shiny Visualisation App}
\usage{
shinyVisApp(con, table_name = "tcga_rnaseq_data")
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{table_name}{The table in the SQLite database containing the expression data.  The table should have fields named
patient_id, gene_id and value where gene_id is an ensembl id.  Default tcga_rnaseq_data}
}
\value{
Launches an interactive Shiny application
}
\description{
Provides an interactive view of Collateral Vulnerability pipeline results
}

