% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisServer.R
\name{shinyVisServer}
\alias{shinyVisServer}
\title{Shiny Visualisation Server}
\usage{
shinyVisServer(input, output, con, table_name = "tcga_rnaseq_data")
}
\arguments{
\item{input}{Shiny input list}

\item{output}{Shiny output list}

\item{con}{A \code{SQLiteConnection} object to the database}

\item{table_name}{The table in the SQLite database containing the expression data.  The table should have fields named
patient_id, gene_id and value where gene_id is an ensembl id.  Default tcga_rnaseq_data}
}
\value{
A shiny server
}
\description{
Creates a shiny server for the interactive view of Collateral Vulnerability pipeline results
}

