% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisUI.R
\name{shinyVisUI}
\alias{shinyVisUI}
\title{Shiny Visualisation UI}
\usage{
shinyVisUI()
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}
}
\value{
A shiny UI
}
\description{
Creates a shiny UI for the interactive view of Collateral Vulnerability pipeline results
}

