% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFlymineDetails.R
\name{tableFlymineDetails}
\alias{tableFlymineDetails}
\title{Flymine details table}
\usage{
tableFlymineDetails(con, input_gene)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{input_gene}{An ensembl gene ID}
}
\value{
A ggplot2 object
}
\description{
Extracts flymine analysis details from the SQLite database for a given gene
}

