% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableGenePairsDetails.R
\name{tableGenePairsDetails}
\alias{tableGenePairsDetails}
\title{Gene Pairs table}
\usage{
tableGenePairsDetails(con, input_genes)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{input_genes}{An vector of ensembl gene IDs}
}
\value{
A data frame
}
\description{
For a given set of input genes, returns gene pairs for testing
}

