% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMutationDetails.R
\name{tableMutationDetails}
\alias{tableMutationDetails}
\title{Mutation details table}
\usage{
tableMutationDetails(con, input_gene)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{input_gene}{An ensembl gene ID}
}
\value{
A data frame
}
\description{
Extracts mutation details from the SQLite database for a given gene
}

