% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableParalogDetails.R
\name{tableParalogDetails}
\alias{tableParalogDetails}
\title{Paralog table}
\usage{
tableParalogDetails(con, input_gene)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{input_gene}{An ensembl gene ID}
}
\value{
A ggplot2 object
}
\description{
Extracts paralog information from the SQLite database for a given gene
}

