% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableWormMineDetails.R
\name{tableWormMineDetails}
\alias{tableWormMineDetails}
\title{WormMine details table}
\usage{
tableWormMineDetails(con, input_gene)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{input_gene}{An ensembl gene ID}
}
\value{
A data frame
}
\description{
Extracts wormmine analysis details from the SQLite database for a given gene
}

