% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countHumanParalogs.R
\name{countHumanParalogs}
\alias{countHumanParalogs}
\title{countHumanParalogs}
\usage{
countHumanParalogs(con, genes, alt_mart = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{genes}{A vector of ENSEMBL gene ids}

\item{alt_mart}{A Mart object created using biomaRt::useMart.  This should specify \code{hsapiens_gene_ensembl}
as the dataset.}
}
\value{
a data frame
}
\description{
This function uses the \pkg{biomaRt} package to extract paralog information for supplied human genes
}
\examples{
\dontrun{
ex_con <- setupSQLite()
all_genes <- getAllHumanGenes(ex_con)
set.seed(20120806)
ex_genes <- sample(all_genes$gene_id, 100)

#count paralogs
ex_df <- countHumanParalogs(ex_con, ex_genes)
ex_df[1:10,1:2]

#view the tables in the database
DBI::dbListTables(ex_con)

#larger set of genes
countHumanParalogs(ex_con, sample(all_genes$gene_id, 2000))[1:10,1:2]

#Specify which Mart to use
my_mart <- biomaRt::useMart("ENSEMBL_MART_ENSEMBL", dataset="hsapiens_gene_ensembl", host="sep2015.archive.ensembl.org")
ex_df <- countHumanParalogs(ex_con, ex_genes, alt_mart=my_mart)
ex_df[1:10,1:2]
}
}

