% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBISEPAnalysis.R
\name{doBISEPAnalysis}
\alias{doBISEPAnalysis}
\title{BISEP Analysis}
\usage{
doBISEPAnalysis(con, genes_n = 1e+05, table_name = "tcga_rnaseq_data",
  log2_transform = TRUE, gene_var_th = 0.2, gene_naprop_th = 0.3,
  seed_val = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{genes_n}{Permits a random subsample of genes to be analysed rather than the full set.  Default 100000 - ie all.}

\item{table_name}{The table in the SQLite database containing the data.  The table should have fields named
patient_id, gene_id and value where gene_id is an ensembl id.  Default tcga_rnaseq_data}

\item{log2_transform}{Boolean operator to determine whether data should be log2 transformed.  Default \code{TRUE}}

\item{gene_var_th}{Genes with very low variance can cause BISEP to fail.  Increasing the gene variance
threshold parameter can filter out these genes.  Default 0.2}

\item{gene_naprop_th}{Genes with a higih proportion of NA values can cause BISEP to fail.  Decreasing
the proportion of NA's filter can filter out these genes.  Default 0.3.}

\item{seed_val}{Setting this parameter ensures that consistent BISEP p-values are returned from run
to run.  See \code{\link{set.seed}} for more details.  Default is NULL - no seed set.}
}
\value{
A data frame
}
\description{
The BISEP analysis function identifies genes with bimodel expression profiles.  See \code{\link[BiSEp]{BiSEp}} documentation.
Note that the version of BISEP used is slightly modified and not that from the BiSEp package -
see \code{\link{BISEP_cv}} for more details.
}

