% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFlyMineAnalysis.R
\name{doFlyMineAnalysis}
\alias{doFlyMineAnalysis}
\title{doFlyMineAnalysis}
\usage{
doFlyMineAnalysis(con, genes)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{genes}{A vector of human ENSEMBL gene ids}
}
\value{
a data frame with the results of the analysis
}
\description{
This function determines the lethality status in D melanogaster of a given human gene
}
\details{
Note: queries the WormMine API using Ruby see http://im-dev.wormbase.org/tools/wormmine/api.do?subtab=ruby
}

