% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRNAseqPlot.R
\name{doRNAseqPlot}
\alias{doRNAseqPlot}
\title{RNAseq plot}
\usage{
doRNAseqPlot(con, input_gene)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{gene_id}{A vector of one or more ENSEMBL gene ids}
}
\value{
A ggplot2 object
}
\description{
View RNAseq data for a given gene
}

