% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doWormMineAnalysis.R
\name{doWormMineAnalysis}
\alias{doWormMineAnalysis}
\title{doWormMineAnalysis}
\usage{
doWormMineAnalysis(con, genes)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{genes}{A vector of ENSEMBL gene ids}
}
\value{
a data frame
}
\description{
This function determines the lethality status in C elegans of a given human gene
}

