% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllHumanGenes.R
\name{getAllHumanGenes}
\alias{getAllHumanGenes}
\title{getAllHumanGenes}
\usage{
getAllHumanGenes(con)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}
}
\value{
a data frame
}
\description{
This function uses the \pkg{ensembldb} package to extract information on all human genes
}

