% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrthologs.R
\name{getOrthologs}
\alias{getOrthologs}
\title{getOrthologs}
\usage{
getOrthologs(con, genes, species = "mmusculus")
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{genes}{A vector of ENSEMBL gene ids}

\item{species}{The ENSEMBL species id eg mmusculus for mouse}
}
\value{
a data frame
}
\description{
This function uses the \pkg{biomaRt} package to get ortholog information for supplied human genes
}

