% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTCGARNAseqData.R
\name{getTCGARNAseqData}
\alias{getTCGARNAseqData}
\title{Get TCGA RNAseq data}
\usage{
getTCGARNAseqData(con, releaseDate, sampletag = "01A", sampleids = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{releaseDate}{GDAC release date.  See \code{\link[RCTGA]{downloadTCGA}} documentation}

\item{sampletag}{Refers to the sample part of the barcode.  Default is 01A.  See \link[=https://wiki.nci.nih.gov/display/TCGA/TCGA+barcode]{TCGA wiki}}

\item{sampleids}{A vector of sample ids to include.  Default is NULL - ie no filtering.  See \link[=https://wiki.nci.nih.gov/display/TCGA/TCGA+barcode]{TCGA wiki}}
}
\value{
a data frame
}
\description{
Downloads, filters and populates the SQLite database with TCGA RNAseq data using functionality from the \pkg{RTCGA}
}

