% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBISEPOutput.R
\name{plotBISEPOutput}
\alias{plotBISEPOutput}
\title{BISEP Plot}
\usage{
plotBISEPOutput(con, highlighted_gene = NULL, pi_value_th = 0.2,
  bisep_pval_th = 0.1)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object}

\item{highlighted_gene}{Highlight the provided gene on the plot.  Default NULL - ie don't highlight anything}

\item{pi_value_th}{Set a PI value threshold.  Default 0.2}

\item{bisep_pval_th}{Set a BISEP p-value threshold.  Default 0.1}
}
\value{
A ggplot2 object
}
\description{
Plot to visualise BISEP output
}

