% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisApp.R
\name{shinyVisApp}
\alias{shinyVisApp}
\title{Shiny Visualisation App}
\usage{
shinyVisApp(con)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}
}
\value{
Launches an interactive Shiny application
}
\description{
Provides an interactive view of Collateral Vulnerability pipeline results
}

