% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisServer.R
\name{shinyVisServer}
\alias{shinyVisServer}
\title{Shiny Visualisation Server}
\usage{
shinyVisServer(input, output, con)
}
\arguments{
\item{input}{Blah}

\item{output}{Blah}

\item{con}{A \code{SQLiteConnection} object to the database}
}
\value{
A shiny server
}
\description{
Creates a shiny server for the interactive view of Collateral Vulnerability pipeline results
}

